Pull CustomSection & Preview out as well
