+++ src/components/CustomSection.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-
 import Grid from "./Grid";
 import Section from "./Section";
