+++ src/components/IndexPageTemplate.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-
 import Header from "./Header";
 import Preview from "./Preview";
 import CustomSection from "./CustomSection";
@@ -13,21 +12,15 @@ const Wrapper = styled.div`
   width: 100%;
 `;
 
-function IndexPageTemplate({ config, children }) {
-  const hasPreview = !!config.preview;
-  const hasCustomSection = !!config.customSection;
-  return (
-    <Wrapper>
-      <Header {...config.header} />
-      {children}
-      {hasPreview ? <Preview {...config.preview} /> : null}
-      {hasCustomSection ? <CustomSection {...config.customSection} /> : null}
-      <GetStarted {...config.getStarted} />
-      <FeaturedOSS />
-      <Footer />
-    </Wrapper>
-  );
-}
+const IndexPageTemplate = ({ config, children }) => (
+  <Wrapper>
+    <Header {...config.header} />
+    {children}
+    <GetStarted {...config.getStarted} />
+    <FeaturedOSS />
+    <Footer />
+  </Wrapper>
+);
 
 IndexPageTemplate.propTypes = {
   config: PropTypes.shape({
